Create count_menu.py
You and three Friends go to a baseball game and zou offer to go to the concession stand for everyone.
They each order one thing, and you do as well. Nachos and Pizza both cost 6 $. A Cheeseburger meal costs 10 $
Water is 4 $ and Coke is 5 $. Tax is 7%

Determine the total cost of ordering four items from the concession stand. If one of your friends orders something that
isn't on the menu, you will order a Coke for them instead.